+++ src/module/audio/AudioBackup.h
@@ -40,7 +40,7 @@ public:
 
 	void start();
 	void stop();
-	int command(ModuleCommand cmd, int param) override;
+	int64 command(ModuleCommand cmd, int64 param) override;
