+++ src/module/audio/AudioSucker.h
@@ -21,7 +21,7 @@ public:
 
 	Port *source;
 
-	int command(ModuleCommand cmd, int param) override;
+	int64 command(ModuleCommand cmd, int64 param) override;
 
 	void set_channels(int channels);
